fixed integer overflow issue and finished challenge 1
changed the internal types from int to long, since the values were to big for stoi to be properly converted.
this fixed an overflow issue with the range coversion.

+++ day-05/include/AlmanacToBlock.h
@@ -14,7 +14,7 @@ private:
 
 public:
     // here we store the different blocks
-    std::vector<int> seeds;
+    std::vector<long> seeds;
     std::vector<std::string> Seed2Soil;
     std::vector<std::string> Soil2Fertilizer;
     std::vector<std::string> Fertilizer2Water;
